+++ jest.config.js
@@ -7,6 +7,8 @@ export default {
 	moduleNameMapper: {
 		"^(\\.{1,2}/.*)\\.js$": "$1",
 		"^@chenglou/pretext$":
-			"<rootDir>/src/utils/__mocks__/pretext-stub.cjs"
+			"<rootDir>/src/utils/__mocks__/pretext-stub.cjs",
+		"^@chenglou/pretext/rich-inline$":
+			"<rootDir>/src/utils/__mocks__/pretext-rich-inline-stub.cjs"
 	}
 };